move combine reducers up a level
